feat(driver): Merge multiple files
Some industrial automation IDEs force the source code to be splitted in
a few files. For example, they might require that types, global
variables and POUs must be defined in separate files.

To make it easier to use the checker for such projects the `-merge`
option was introduced. It simply merges the content of all input files
to a single temporary file before running analysis.

Resolves #20